Improve sort on table
